users/chatbot/rag_py/requirements.txt
google-generativeai
python-dotenv
chromadb
pysqlite3-binary
